Treat object and dynamic as the same in anonymous function conversions.
This causes code such as `Action<dynamic> a = (object x) => {}` to be allowed. I can't justify this in the spec, but both csc and mcs allow it.

I also added some tests for this conversion in normal method group conversions. Those tests ended up passing, but they probably don't hurt to have.