Two failing tests involving type inference when a type is implicitly convertible both to and from another type.
Note that the List<T>case is supposed to resolve without an error, but the IEnumerable<T> case is supposed to give an error. I do not understand why it is so (perhaps something with generic variance; though I thought that only involved identity conversions), but the behavior is the same in both mcs and csc.