Ported LockThisAnalyzer with Roslyn  using the following An helper method to validate that the lockstatement does contain, or not, a this expression syntax An helper method to validate that a attribute is of type MethodImpl and that the user accesses the enum value of Synchronized in MethodImplOptions
For now, most of the tests are failing because of the text span that I'm giving but I don't know what is wrong and I would need some help with it.

The test TestMethodHasSynchronized gives me an headache.
Breaks my code as when using | does not allow any simple member access expression. I would need to use a BinaryExpressionSyntax to get access to the left/right members. Do I need to make a big recursion code for those cases or just have a simple binary that I would then break in its members to get my hands on the member access expressions and then follow the same logic that I had

As for the code fix provider, I don't really know how it is realy done. I have started to modify it, but I would need some guidance to know how to do them in the future.