ELiminated KeyValuePair from previous analyzer
The TryDiagnostics method was modified in a way to directly evaluate the this expression within a lock statement with the helper boolean method ValidateLockStatementUsesThisExpression.

In a situation where the arguments in the attribute are delimited with a binary expresison syntax,  the recursion method FindMemberAccessExpressionLocationWithinBinaryExpression was developped to find the location a member access expression.

IsSynchronizedEnumValueBeingUsed boolean method was created to make sure of the enum and the value of the enum inside the member access expression.