Fix dirty builds all the time from doc file
The previous path of the documentation file, pointed to bin
caused MSBuild on Windows to always build the output assembly
again every time. By moving the doc file to the intermediate
output path, the problem goes away and MSBuild properly
detects that the project is up-to-date.

The doc file is automatically copied to the output path, so
no behavior changes from the previous value.